Small fixes to index management
